backend/requirements.txt
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
Authlib==1.2.1
psycopg2-binary==2.9.7
